feat(Liquid.Core): update Liquid.Core references. This feature is a partial refactoring of Liquid.Core, with main objectives to make this package net6 compatible, change from Newtonsoft.Json to System.Text.Json, separate extension methods from helper methods. Implemented issues #130 and #131
