+++ src/Liquid.Core/Extensions/DateTimeExtension.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Globalization;
 
-namespace Liquid.Core.Utils
+namespace Liquid.Core.Extensions
 {
     /// <summary>
     /// Date time extensions class.
     /// </summary>
-    public static class DateTimeUtils
+    public static class DateTimeExtension
     {
         /// <summary>
         /// Converts datetime to Iso 8601 format.
