+++ src/Liquid.Core/Extensions/IntExtension.cs
@@ -5,7 +5,7 @@ namespace Liquid.Core.Utils
     /// <summary>
     /// Number Extensions Class.
     /// </summary>
-    public static class NumberUtils
+    public static class IntExtension
     {
         /// <summary>
         /// Determines whether a number is is prime number.
